# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API สำหรับจัดการกับระบบสูตรอาหาร",
        "title": "Wongnok API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/auth/callback": {
            "get": {
                "description": "Keycloak redirect มาที่นี่เองอัตโนมัติ ไม่ใช่สิ่งที่ frontend เรียกตรงๆ",
                "tags": [
                    "auth"
                ],
                "summary": "รับ callback จาก Keycloak",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization code",
                        "name": "code",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "State ที่ตรงกับตอน /auth/login",
                        "name": "state",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect กลับ frontend พร้อม ticket",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/wongnok_internal_httputil.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/wongnok_internal_httputil.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/wongnok_internal_httputil.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/exchange": {
            "post": {
                "description": "Frontend เรียกผ่าน axios หลังถูก redirect กลับมาพร้อม ticket",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "แลก ticket เป็น credential จริง",
                "parameters": [
                    {
                        "description": "ticket",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_auth.ExchangeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_auth.Credential"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/wongnok_internal_httputil.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/wongnok_internal_httputil.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "get": {
                "description": "Redirect ไปหน้า login ของ Keycloak",
                "tags": [
                    "auth"
                ],
                "summary": "เริ่ม login ผ่าน Keycloak",
                "responses": {
                    "302": {
                        "description": "Redirect ไป Keycloak",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/wongnok_internal_httputil.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Revoke refresh token ที่ Keycloak จริง",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "ออกจากระบบ",
                "parameters": [
                    {
                        "description": "refresh token",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_auth.LogoutRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Logout สำเร็จ"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/wongnok_internal_httputil.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/wongnok_internal_httputil.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "สร้าง user โดยรับข้อมูลมาจาก Client และ generate uuid ให้ พร้อมส่งรายละเอียดของ user ที่ถูกสร้างกลับไป",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "สร้าง user",
                "parameters": [
                    {
                        "description": "รายละเอียดสำหรับสร้าง User",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_user.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_user.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/wongnok_internal_httputil.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/wongnok_internal_httputil.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "ค้นหาข้อมูล User จาก UUID แล้วคืนข้อมูล User ที่เจอกลับมา",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "ดีงข้อมูลจาก user แบบรายคน",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "User ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_user.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/wongnok_internal_httputil.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/wongnok_internal_httputil.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/wongnok_internal_httputil.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "internal_auth.Credential": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string"
                },
                "expiresAt": {
                    "type": "string"
                },
                "refreshToken": {
                    "type": "string"
                },
                "tokenType": {
                    "type": "string"
                }
            }
        },
        "internal_auth.ExchangeRequest": {
            "type": "object",
            "required": [
                "ticket"
            ],
            "properties": {
                "ticket": {
                    "type": "string",
                    "example": "pQx7...base64url..."
                }
            }
        },
        "internal_auth.LogoutRequest": {
            "type": "object",
            "required": [
                "refreshToken"
            ],
            "properties": {
                "refreshToken": {
                    "type": "string",
                    "example": "eyJhbGci..."
                }
            }
        },
        "internal_user.CreateUserRequest": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "somchai@pea.co.th"
                },
                "name": {
                    "type": "string",
                    "example": "สมชาย ใจดี"
                }
            }
        },
        "internal_user.UserResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "somchai@pea.co.th"
                },
                "id": {
                    "type": "string",
                    "example": "eb238dfe-9d49-4345-9cc9-350c50b429e6"
                },
                "name": {
                    "type": "string",
                    "example": "สมชาย ใจดี"
                }
            }
        },
        "wongnok_internal_httputil.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "พิมพ์ \"Bearer\" ตามด้วย space แล้วตามด้วย JWT token เช่น \"Bearer eyJhbGci...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}